+++ src/main/java/cart/auth/MemberArgumentResolver.java
@@ -1,10 +1,9 @@
-package cart.ui;
+package cart.auth;
 
 import cart.dao.AuthDao;
 import cart.domain.Member;
 import cart.entity.MemberEntity;
 import cart.exception.AuthenticationException;
-import java.util.NoSuchElementException;
 import org.apache.tomcat.util.codec.binary.Base64;
 import org.springframework.core.MethodParameter;
 import org.springframework.http.HttpHeaders;
@@ -27,7 +26,7 @@ public class MemberArgumentResolver implements HandlerMethodArgumentResolver {
     }
 
     @Override
-    public Object resolveArgument(final MethodParameter parameter, final ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
+    public Object resolveArgument(final MethodParameter parameter, final ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         String authorization = webRequest.getHeader(HttpHeaders.AUTHORIZATION);
         if (authorization == null) {
             return null;
