+++ src/main/java/cart/config/WebMvcConfig.java
@@ -1,7 +1,7 @@
-package cart;
+package cart.config;
 
 import cart.dao.AuthDao;
-import cart.ui.MemberArgumentResolver;
+import cart.auth.MemberArgumentResolver;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
@@ -11,7 +11,6 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 public class WebMvcConfig implements WebMvcConfigurer {
 
     private final AuthDao authDao;
-
 
     public WebMvcConfig(final AuthDao authDao) {
         this.authDao = authDao;
